requirements.txt
Django==2.0.7
django-debug-toolbar==1.9.1
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-shortuuidfield==0.1.3
django-simple-captcha==0.5.9
djangorestframework==3.8.2
mysqlclient==1.3.7
Pillow==5.2.0
pymemcache==1.4.4
python-memcached==1.59
pytz==2018.5
shortuuid==0.5.0
six==1.11.0
sqlparse==0.2.4
